Update for ExToolbox
Adds:
- Entity now has `get_id()` function
- Save File dialog now automatically appends default extension in file name if user doesn't specify
Changes:
- SceneCamera Orthographic Camera now automatically resizes itself to match viewport size
- Scene hierarchy now won't crash the program anymore when user add an existed component
- Scene hierarchy now uses strncpy_s() instead of std::strncpy() (fix some warnings)
- Remove redundancy struct `vec2` in `Input` class
- Entity `get_component()` and `has_component()` are now const
- Cleanup `FileDialogs` functions

+++ ExWorkshop/src/exw/maths/math_utils.h
@@ -12,7 +12,7 @@ namespace exw::maths
 {
     namespace
     {
-        constexpr float Pi = 3.1415926535897932;
+        constexpr float Pi = (float)3.1415926535897932;
         constexpr float Pi_over_180d = Pi / 180.0f;
         constexpr float Pi_divides_180d = 180.0f / Pi;
